frontend/requirements.txt
streamlit==1.46.0
pandas>=2.0.0
plotly>=5.15.0
wordcloud>=1.9.2
textblob>=0.17.1
networkx>=3.1
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
